Use int* typedefs in some more places.
Especially important in simd classes, where we have to be sure that size is always 16 and 32 bit.


git-svn-id: https://piranha-psm.googlecode.com/svn/trunk@642 f32a13d4-301a-0410-ba96-7b22a51b2342

+++ src/trig_simd_array.h
@@ -67,7 +67,7 @@ namespace piranha
         size_t i;
         for (i=0;i<boost::minmax((size_t)d,w-1).get<0>();++i)
         {
-          private_container_[i]=utils::lexical_converter<int>(sd[i]);
+          private_container_[i]=utils::lexical_converter<int16>(sd[i]);
         }
         for (;i<d;++i)
         {
